*Improved the parameter validation of several functions and made minor style changes to some older code. *Changed the behavior of the logic that reads the pro guitar/bass tunings from song.ini.  These values will not override those in the project when a project is loaded, as Rocksmith supports 4 or 6 string counts to be authored in either track type. *Added the CTRL+W shortcut key to "Note>Star power>Mark". *Changed eof_pos_distance() to use double precision floating point.  It's only used in the grid snap logic, but even a negligible accuracy improvement is desirable. *Corrected the variable type of the "note" parameter in eof_track_add_create_note() to reflect that it is a note bitmask and not a note index. *Minor changes to satisfy Splint.

+++ src/editor.h
@@ -130,7 +130,7 @@ unsigned char eof_find_pen_note_mask(void);
 
 void eof_editor_logic_common(void);
 	//Performs editor logic common to both the editor and vocal editor windows (some variable initialization, handling beat marker, seek bar and playback controls)
-float eof_pos_distance(float p1, float p2);
+double eof_pos_distance(double p1, double p2);
 	//Returns the difference between the two floating point numbers as a positive value
 void eof_get_snap_ts(EOF_SNAP_DATA * sp, int beat);
 	//Finds the time signature in effect for the specified beat and stores the numerator and denominator into sp
